refactor(spice): block-level finalize owns commit + post-commit ops
`apply_block_postprocessing` now absorbs `update_flat_storage_head` and
`gc_memtrie_roots` from the actor. The helper owns the entire block-level
finalize: spice-head writes + commit + flat-storage advance + memtrie GC.
Doc comment explains the pre-commit vs post-commit boundary and why this
helper differs from `apply_chunk_postprocessing` (which takes
`&mut StoreUpdate` and lets caller commit). The actor's `finalize_block`
collapses to a thin wrapper.

Also drops the `advance_execution_head` call from `process_apply_chunk_results`:
the just-finalized block's head advance already happens via
`finalize_block(block_hash)`, and the canonical-next walk can't reach
that block from the prior head under forks anyway. The walk still runs
on `start_actor` for restart-after-crash backlog.

Reverts the `is_descendant_of_final_execution_head` defensive
DBNotFoundErr fix from the prior commit — with the head-advancement
chain restored to normal, the walk stays short and the latent
`.unwrap()` doesn't fire.
